Revert "Feat style"
