+++ .babelrc
@@ -1,3 +1,4 @@
 {
+  "presets": ["env"],
   "plugins": [["transform-react-jsx", { "pragma": "h" }]]
 }
