+++ package.json
@@ -23,8 +23,8 @@
   "homepage": "https://github.com/Art2B/fr-canvas#readme",
   "devDependencies": {
     "babel-plugin-transform-react-jsx": "^6.24.1",
-    "node-sass": "^4.7.2",
-    "parcel-bundler": "^1.6.0",
+    "babel-preset-env": "^1.6.1",
+    "parcel-bundler": "^1.5.1",
     "standard": "^10.0.3"
   },
   "dependencies": {
