--- src/assets/styles/components/_footer.scss
@@ -1,14 +1,0 @@
-footer {
-  text-align: center;
-  width: 100%;
-  padding: 10px 0;
-  position: fixed;
-  bottom: 0;
-  .border {
-    width: 40%;
-    height: 1px;
-    display: block;
-    background: #292824;
-    margin: 0px auto 10px;
-  }
-}
