--- src/assets/styles/components/_header.scss
@@ -1,13 +1,0 @@
-header {
-  text-align: center;
-  margin: 10px auto 40px;
-  .hexagon-anim {
-    height: 100px;
-    width: 100px;
-    border: 2px solid #292824;
-    margin: 0 auto 10px;
-  }
-  h1 {
-    font-size: 32px;
-  }
-}
