--- src/assets/styles/components/_layout.scss
@@ -1,30 +1,0 @@
-html, body {
-  margin: 0;
-  padding: 0;
-
-  min-height: 100%;
-  height: 100%;
-  max-height: 100%;
-  background-color: $bg-color;
-}
-
-.app-view {
-  height: 100%;
-  padding: 10px 0;
-  box-sizing: border-box;
-  font-family: Courier New;
-  color: $primary-color;
-}
-
-.explanations {
-  width: 90%;
-  margin: auto;
-}
-
-a {
-  color: $secondary-color;
-  text-decoration: none;
-  &:hover, &:visited, &:active {
-    color: $secondary-color;
-  }
-}
