--- src/assets/styles/components/_main.scss
@@ -1,44 +1,0 @@
-main {
-  width: 90%;
-  min-height: 50vw;
-  margin: 20px auto 60px;
-  border: 4px solid $primary-color;
-  position: relative;
-  .file-input {
-    position: absolute;
-    z-index: 5;
-    height: 40vw;
-    width: 40vw;
-    top: 0; bottom: 0;
-    left: 0; right: 0;
-    margin: auto;
-    cursor: pointer;
-    text-align: center;
-    line-height: 40vw;
-    color: $secondary-color;
-
-    /*Vertical*/
-    background-image: linear-gradient($primary-color 33%, $bg-color 0%);
-    background-position: right;
-    background-size: 1px 4px;
-    .text-bg {
-      background: $bg-color;
-      padding: 2px 5px;
-    }
-  }
-  input {
-    display:none;
-  }
-  canvas {
-    width: 100%;
-    height: 100%;
-    position: relative;
-    z-index: 1;
-  }
-
-  &.is-file {
-    .file-input {
-      display: none;
-    }
-  }
-}
