--- src/assets/styles/config/_variables.scss
@@ -1,3 +1,0 @@
-$bg-color: #FFFBE9;
-$primary-color: #292824;
-$secondary-color: indianred;
