--- src/assets/styles/index.scss
@@ -1,8 +1,0 @@
-@import 'lib/reset';
-
-@import 'config/variables';
-
-@import 'components/layout';
-@import 'components/header';
-@import 'components/main';
-@import 'components/footer';
