+++ src/assets/styles/style.css
@@ -1,0 +1,4 @@
+canvas {
+  border: 2px solid black;
+  margin: 10px;
+}
