+++ src/index.html
@@ -3,9 +3,9 @@
 <head>
   <meta charset="UTF-8">
   <title>Canvas workaround</title>
+  <link rel="stylesheet" href="assets/styles/style.css">
 </head>
 <body>
   <script src="scripts/index.js"></script>
-  <script defer src="https://use.fontawesome.com/releases/v5.0.4/js/all.js"></script>
 </body>
 </html>
