+++ src/scripts/index.js
@@ -1,10 +1,6 @@
 import 'babel-polyfill'
 import { h, app } from 'hyperapp'
 
-import Hexagon from './hexagon'
-
-import './../assets/styles/index.scss'
-
 const state = {
   image: ''
 }
@@ -32,11 +28,7 @@ const drawImage = (canvas, base64) => {
     canvas.width = img.width
     canvas.height = img.height
     // Draw image on canvas
-    ctx.filter = 'grayscale(100%)'
     ctx.drawImage(img, 0, 0)
-    // Draw hexagons on image
-    const hexa = new Hexagon(canvas, 100)
-    hexa.draw()
   })
   img.src = base64
 }
@@ -50,23 +42,9 @@ const Canvas = ({ image }) => (
 )
 
 const view = (state, actions) => (
-  <div className='app-view'>
-    <header>
-      <div className='hexagon-anim' />
-      <h1>Hexag on canvas</h1>
-    </header>
-    <section className='explanations'>
-      <p>Choose an image, let me process it, and BAM! You have a cool image made with hexagons.</p>
-    </section>
-    <main className={state.image ? 'is-file' : ''}>
-      <label className='file-input' for='fr-upload'><span className='text-bg'>Choose a <i A className="fas fa-fire" /> af image.</span></label>
-      <input type='file' id='fr-upload' size='50' onchange={e => actions.handleFile(e.target.files[0])} />
-      <Canvas image={state.image} />
-    </main>
-    <footer>
-      <div className='border'/>
-      Made with <i className='far fa-keyboard' /> by <a href='https://twitter.com/mr_marabout' target='_blank'>Art2B</a>
-    </footer>
+  <div>
+    <input type='file' id='fr-upload' multiple size='50' onchange={e => actions.handleFile(e.target.files[0])} />
+    <Canvas image={state.image} />
   </div>
 )
